Moved common types #205
