+++ lisa/lisa-analyses/src/test/java/it/unive/lisa/analysis/types/InferredTypesTest.java
@@ -12,6 +12,10 @@ import it.unive.lisa.program.Program;
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CodeLocation;
 import it.unive.lisa.program.cfg.ProgramPoint;
+import it.unive.lisa.program.type.BoolType;
+import it.unive.lisa.program.type.Float32Type;
+import it.unive.lisa.program.type.Int32Type;
+import it.unive.lisa.program.type.StringType;
 import it.unive.lisa.symbolic.value.operator.binary.BinaryOperator;
 import it.unive.lisa.symbolic.value.operator.binary.ComparisonEq;
 import it.unive.lisa.symbolic.value.operator.binary.ComparisonGe;
@@ -48,10 +52,6 @@ import it.unive.lisa.type.Type;
 import it.unive.lisa.type.TypeSystem;
 import it.unive.lisa.type.TypeTokenType;
 import it.unive.lisa.type.Untyped;
-import it.unive.lisa.type.common.BoolType;
-import it.unive.lisa.type.common.Float32Type;
-import it.unive.lisa.type.common.Int32Type;
-import it.unive.lisa.type.common.StringType;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
