+++ lisa/lisa-analyses/src/test/java/it/unive/lisa/analysis/value/SubstitutionTest.java
@@ -12,10 +12,10 @@ import it.unive.lisa.program.SyntheticLocation;
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CodeLocation;
 import it.unive.lisa.program.cfg.ProgramPoint;
+import it.unive.lisa.program.type.Int32Type;
 import it.unive.lisa.symbolic.value.Identifier;
 import it.unive.lisa.symbolic.value.ValueExpression;
 import it.unive.lisa.symbolic.value.Variable;
-import it.unive.lisa.type.common.Int32Type;
 import it.unive.lisa.util.collections.CollectionsDiffBuilder;
 import java.util.ArrayList;
 import java.util.Arrays;
