+++ lisa/lisa-imp/src/main/java/it/unive/lisa/imp/constructs/StringLength.java
@@ -11,8 +11,8 @@ import it.unive.lisa.program.cfg.statement.Expression;
 import it.unive.lisa.program.cfg.statement.PluggableStatement;
 import it.unive.lisa.program.cfg.statement.Statement;
 import it.unive.lisa.program.cfg.statement.string.Length;
-import it.unive.lisa.type.common.Int32Type;
-import it.unive.lisa.type.common.StringType;
+import it.unive.lisa.program.type.Int32Type;
+import it.unive.lisa.program.type.StringType;
 
 /**
  * The native construct representing the length operation. This construct can be
