+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/cfg/statement/literal/Int16Literal.java
@@ -2,7 +2,7 @@ package it.unive.lisa.program.cfg.statement.literal;
 
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CodeLocation;
-import it.unive.lisa.type.common.Int16Type;
+import it.unive.lisa.program.type.Int16Type;
 
 /**
  * A 16-bit signed {@link Literal} representing a constant integral value.
