+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/cfg/statement/literal/Int32Literal.java
@@ -2,7 +2,7 @@ package it.unive.lisa.program.cfg.statement.literal;
 
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CodeLocation;
-import it.unive.lisa.type.common.Int32Type;
+import it.unive.lisa.program.type.Int32Type;
 
 /**
  * A 32-bit signed {@link Literal} representing a constant integral value.
