+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/cfg/statement/literal/UInt8Literal.java
@@ -2,7 +2,7 @@ package it.unive.lisa.program.cfg.statement.literal;
 
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.CodeLocation;
-import it.unive.lisa.type.common.UInt8Type;
+import it.unive.lisa.program.type.UInt8Type;
 
 /**
  * An 8-bit unsigned {@link Literal} representing a constant integral value.
