+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/type/BoolType.java
@@ -1,5 +1,6 @@
-package it.unive.lisa.type.common;
+package it.unive.lisa.program.type;
 
+import it.unive.lisa.type.BooleanType;
 import it.unive.lisa.type.Type;
 import it.unive.lisa.type.TypeSystem;
 import it.unive.lisa.type.Untyped;
@@ -7,53 +8,52 @@ import java.util.Collections;
 import java.util.Set;
 
 /**
- * An implementation of the {@link it.unive.lisa.type.StringType}. The only
- * singleton instance of this class can be retrieved trough field
- * {@link #INSTANCE}. <br>
+ * An internal implementation of the {@link BooleanType}. The only singleton
+ * instance of this class can be retrieved trough field {@link #INSTANCE}. <br>
  * <br>
  * Instances of this class are equal to all other classes that implement the
- * {@link it.unive.lisa.type.StringType}.
+ * {@link BooleanType}.
  * 
  * @author <a href="mailto:luca.negrini@unive.it">Luca Negrini</a>
  */
-public class StringType implements it.unive.lisa.type.StringType {
+public class BoolType implements BooleanType {
 
 	/**
 	 * The singleton instance of this class.
 	 */
-	public static final StringType INSTANCE = new StringType();
+	public static final BoolType INSTANCE = new BoolType();
 
 	/**
 	 * Builds the type. This constructor is visible to allow subclassing:
 	 * instances of this class should be unique, and the singleton can be
 	 * retrieved through field {@link #INSTANCE}.
 	 */
-	protected StringType() {
+	protected BoolType() {
 	}
 
 	@Override
 	public boolean canBeAssignedTo(Type other) {
-		return other.isStringType() || other.isUntyped();
+		return other.isBooleanType() || other.isUntyped();
 	}
 
 	@Override
 	public Type commonSupertype(Type other) {
-		return other.isStringType() ? this : Untyped.INSTANCE;
+		return other.isBooleanType() ? this : Untyped.INSTANCE;
 	}
 
 	@Override
 	public String toString() {
-		return "string";
+		return "bool";
 	}
 
 	@Override
 	public boolean equals(Object other) {
-		return other instanceof it.unive.lisa.type.StringType;
+		return other instanceof BooleanType;
 	}
 
 	@Override
 	public int hashCode() {
-		return it.unive.lisa.type.StringType.class.getName().hashCode();
+		return BooleanType.class.getName().hashCode();
 	}
 
 	@Override
