+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/type/UInt32Type.java
@@ -1,4 +1,4 @@
-package it.unive.lisa.type.common;
+package it.unive.lisa.program.type;
 
 import it.unive.lisa.type.NumericType;
 import it.unive.lisa.type.Type;
@@ -8,16 +8,16 @@ import java.util.Collections;
 import java.util.Set;
 
 /**
- * A signed 64-bit integral {@link NumericType}. The only singleton instance of
- * this class can be retrieved trough field {@link #INSTANCE}.<br>
+ * An unsigned 32-bit integral {@link NumericType}. The only singleton instance
+ * of this class can be retrieved trough field {@link #INSTANCE}.<br>
  * <br>
  * Instances of this class are equal to all other classes that implement the
  * {@link NumericType} interface, and for which {@link #isIntegral()} and
- * {@link #is64Bits()} yield {@code true}. An instance of Int64 is assumed to be
+ * {@link #is32Bits()} yield {@code true}. An instance of Int32 is assumed to be
  * assignable to any {@link NumericType}, with possible loss of information.
  * <br>
  * <br>
- * The common supertype between an Int64 instance {@code t1} and another type
+ * The common supertype between an Int32 instance {@code t1} and another type
  * instance {@code t2} is {@link Untyped} if {@code t2} is not a
  * {@link NumericType}. Otherwise, the supertype is chosen according to
  * {@link NumericType#supertype(NumericType)}. <br>
@@ -27,19 +27,19 @@ import java.util.Set;
  * 
  * @author <a href="mailto:luca.negrini@unive.it">Luca Negrini</a>
  */
-public class Int64Type implements NumericType {
+public class UInt32Type implements NumericType {
 
 	/**
 	 * The unique singleton instance of this type.
 	 */
-	public static final Int64Type INSTANCE = new Int64Type();
+	public static final UInt32Type INSTANCE = new UInt32Type();
 
 	/**
 	 * Builds the type. This constructor is visible to allow subclassing:
 	 * instances of this class should be unique, and the singleton can be
 	 * retrieved through field {@link #INSTANCE}.
 	 */
-	protected Int64Type() {
+	protected UInt32Type() {
 	}
 
 	@Override
@@ -54,17 +54,17 @@ public class Int64Type implements NumericType {
 
 	@Override
 	public boolean is32Bits() {
-		return false;
-	}
-
-	@Override
-	public boolean is64Bits() {
 		return true;
 	}
 
 	@Override
-	public boolean isUnsigned() {
+	public boolean is64Bits() {
 		return false;
+	}
+
+	@Override
+	public boolean isUnsigned() {
+		return true;
 	}
 
 	@Override
@@ -87,7 +87,7 @@ public class Int64Type implements NumericType {
 
 	@Override
 	public String toString() {
-		return "int64";
+		return "int32";
 	}
 
 	@Override
@@ -100,7 +100,7 @@ public class Int64Type implements NumericType {
 
 	@Override
 	public int hashCode() {
-		return Int64Type.class.getName().hashCode();
+		return UInt32Type.class.getName().hashCode();
 	}
 
 	@Override
