+++ lisa/lisa-program/src/main/java/it/unive/lisa/program/type/UInt64Type.java
@@ -1,4 +1,4 @@
-package it.unive.lisa.type.common;
+package it.unive.lisa.program.type;
 
 import it.unive.lisa.type.NumericType;
 import it.unive.lisa.type.Type;
@@ -8,16 +8,16 @@ import java.util.Collections;
 import java.util.Set;
 
 /**
- * An unsigned 8-bit integral {@link NumericType}. The only singleton instance
+ * An unsigned 64-bit integral {@link NumericType}. The only singleton instance
  * of this class can be retrieved trough field {@link #INSTANCE}.<br>
  * <br>
  * Instances of this class are equal to all other classes that implement the
  * {@link NumericType} interface, and for which {@link #isIntegral()} and
- * {@link #is8Bits()} yield {@code true}. An instance of Int8 is assumed to be
+ * {@link #is64Bits()} yield {@code true}. An instance of Int64 is assumed to be
  * assignable to any {@link NumericType}, with possible loss of information.
  * <br>
  * <br>
- * The common supertype between an Int8 instance {@code t1} and another type
+ * The common supertype between an Int64 instance {@code t1} and another type
  * instance {@code t2} is {@link Untyped} if {@code t2} is not a
  * {@link NumericType}. Otherwise, the supertype is chosen according to
  * {@link NumericType#supertype(NumericType)}. <br>
@@ -27,24 +27,24 @@ import java.util.Set;
  * 
  * @author <a href="mailto:luca.negrini@unive.it">Luca Negrini</a>
  */
-public class UInt8Type implements NumericType {
+public class UInt64Type implements NumericType {
 
 	/**
 	 * The unique singleton instance of this type.
 	 */
-	public static final UInt8Type INSTANCE = new UInt8Type();
+	public static final UInt64Type INSTANCE = new UInt64Type();
 
 	/**
 	 * Builds the type. This constructor is visible to allow subclassing:
 	 * instances of this class should be unique, and the singleton can be
 	 * retrieved through field {@link #INSTANCE}.
 	 */
-	protected UInt8Type() {
+	protected UInt64Type() {
 	}
 
 	@Override
 	public boolean is8Bits() {
-		return true;
+		return false;
 	}
 
 	@Override
@@ -59,7 +59,7 @@ public class UInt8Type implements NumericType {
 
 	@Override
 	public boolean is64Bits() {
-		return false;
+		return true;
 	}
 
 	@Override
@@ -87,7 +87,7 @@ public class UInt8Type implements NumericType {
 
 	@Override
 	public String toString() {
-		return "int8";
+		return "int64";
 	}
 
 	@Override
@@ -100,7 +100,7 @@ public class UInt8Type implements NumericType {
 
 	@Override
 	public int hashCode() {
-		return UInt8Type.class.getName().hashCode();
+		return UInt64Type.class.getName().hashCode();
 	}
 
 	@Override
